lock exchange: move diagnostics and plotting to separate files
- plotting callback depends on matplotlib, reverts to no op if missing
